# Conflicts:
#	hyjf_cs_server/hyjf_cs_trade/src/main/java/com/hyjf/cs/trade/client/AmBorrowClient.java
#	hyjf_cs_server/hyjf_cs_trade/src/main/java/com/hyjf/cs/trade/client/AmUserClient.java
#	hyjf_cs_server/hyjf_cs_trade/src/main/java/com/hyjf/cs/trade/client/impl/AmBorrowClientImpl.java
#	hyjf_cs_server/hyjf_cs_trade/src/main/java/com/hyjf/cs/trade/client/impl/AmUserClientImpl.java

+++ hyjf_am_server/hyjf_am_core/hyjf_am_user/src/main/java/com/hyjf/am/user/dao/model/auto/UserEvalation.java
@@ -12,7 +12,7 @@ public class UserEvalation implements Serializable {
 
     private Integer answerId;
 
-    private Byte sort;
+    private Integer sort;
 
     private Date createTime;
 
@@ -52,11 +52,11 @@ public class UserEvalation implements Serializable {
         this.answerId = answerId;
     }
 
-    public Byte getSort() {
+    public Integer getSort() {
         return sort;
     }
 
-    public void setSort(Byte sort) {
+    public void setSort(Integer sort) {
         this.sort = sort;
     }
